Removed try/catch for returning the status of variable setting names
Since try/catch is very slow for this purpose, I've removed it. LiveSplit will still throw warnings to DbgView, but it should not halt execution or affect functionality. Future versions of LiveSplit, when they get released, should offer a ContainsKey method on the settings object that will allow me to explicitly check if a setting exists.